+++ src/main/java/commands/DeleteCommand.java
@@ -1,19 +1,27 @@
-public class MarkCommand extends Command {
+package commands;
+
+import core.Task;
+import core.TaskList;
+import core.Ui;
+import util.Storage;
+import util.TronaldDumpException;
+
+public class DeleteCommand extends Command {
     
-    public MarkCommand(TaskList taskList, Storage storage) {
+    public DeleteCommand(TaskList taskList, Storage storage) {
         super(taskList, storage);
     }
     
     @Override
     public void execute(String input, String[] parts) throws TronaldDumpException {
         if (parts.length != 2) {
-            throw new TronaldDumpException("Please specify the task number to mark.");
+            throw new TronaldDumpException("Please specify the task number to delete.");
         }
         
         int taskNumber = Integer.parseInt(parts[1]) - 1;
         if (taskNumber >= 0 && taskNumber < taskList.getSize()) {
-            taskList.markTask(taskNumber);
-            Ui.showMarkTask(taskList, taskNumber);
+            Task removedTask = taskList.removeTask(taskNumber);
+            Ui.showDeleteTask(taskList, removedTask, taskNumber);
         } else {
             throw new TronaldDumpException("Invalid task number.");
         }
