+++ src/main/java/commands/MarkCommand.java
@@ -1,21 +1,26 @@
+package commands;
 
+import core.TaskList;
+import core.Ui;
+import util.Storage;
+import util.TronaldDumpException;
 
-public class UnmarkCommand extends Command {
+public class MarkCommand extends Command {
     
-    public UnmarkCommand(TaskList taskList, Storage storage) {
+    public MarkCommand(TaskList taskList, Storage storage) {
         super(taskList, storage);
     }
     
     @Override
     public void execute(String input, String[] parts) throws TronaldDumpException {
         if (parts.length != 2) {
-            throw new TronaldDumpException("Please specify the task number to unmark.");
+            throw new TronaldDumpException("Please specify the task number to mark.");
         }
         
         int taskNumber = Integer.parseInt(parts[1]) - 1;
         if (taskNumber >= 0 && taskNumber < taskList.getSize()) {
-            taskList.unmarkTask(taskNumber);
-            Ui.showUnmarkTask(taskList, taskNumber);
+            taskList.markTask(taskNumber);
+            Ui.showMarkTask(taskList, taskNumber);
         } else {
             throw new TronaldDumpException("Invalid task number.");
         }
