+++ src/main/java/core/Event.java
@@ -1,3 +1,7 @@
+package core;
+
+import util.TronaldDumpException;
+
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
