+++ src/main/java/core/Task.java
@@ -1,3 +1,5 @@
+package core;
+
 public abstract class Task {
     private String description;
     private boolean isDone;
