+++ src/main/java/core/Todo.java
@@ -1,3 +1,5 @@
+package core;
+
 public class Todo extends Task{
 
     public Todo(String description, boolean isDone) {
