+++ src/main/java/util/Parser.java
@@ -1,3 +1,5 @@
+package util;
+
 public class Parser {
     public static String[] parse(String input) {
         return input.trim().split("\\s+");
