+++ src/main/java/util/TronaldDumpException.java
@@ -1,3 +1,5 @@
+package util;
+
 public class TronaldDumpException extends RuntimeException {
 
     public TronaldDumpException(String message) {
