+++ text-ui-test/runtest.sh
@@ -6,10 +6,13 @@ then
     mkdir ../bin
 fi
 
+# remove old compiled classes
+rm -rf ../bin/*
+
 # delete output from previous run
 if [ -e "./ACTUAL.TXT" ]
 then
-    rm ACTUAL.TXT
+    rm ./ACTUAL.TXT
 fi
 
 if [ -e "./data/tronalddump.txt" ]
@@ -17,22 +20,26 @@ then
     rm ./data/tronalddump.txt
 fi
 
+
 # compile the code into the bin folder, terminates if error occurred
-if ! javac -cp ../src/main/java -Xlint:none -d ../bin ../src/main/java/*.java
+SRC=../src/main/java
+BIN=../bin
+find "$SRC" -name "*.java" > sources.txt
+if ! javac -Xlint:none -d "$BIN" @sources.txt
 then
     echo "********** BUILD FAILURE **********"
     exit 1
 fi
 
 # run the program, feed commands from input.txt file and redirect the output to the ACTUAL.TXT
-java -classpath ../bin TronaldDump < input2.txt > ACTUAL.TXT
+java -classpath ../bin core.TronaldDump < input3.txt > ACTUAL.TXT
 
 # convert to UNIX format
-cp EXPECTED2.TXT EXPECTED2-UNIX.TXT
-dos2unix ACTUAL.TXT EXPECTED2-UNIX.TXT
+cp EXPECTED3.TXT EXPECTED3-UNIX.TXT
+dos2unix ACTUAL.TXT EXPECTED3-UNIX.TXT
 
 # compare the output to the expected output
-diff ACTUAL.TXT EXPECTED2-UNIX.TXT
+diff ACTUAL.TXT EXPECTED3-UNIX.TXT
 if [ $? -eq 0 ]
 then
     echo "Test result: PASSED"
