+++ text-ui-test/sources.txt
@@ -1,0 +1,20 @@
+../src/main/java/core/TaskList.java
+../src/main/java/core/Ui.java
+../src/main/java/core/TronaldDump.java
+../src/main/java/core/Event.java
+../src/main/java/core/Task.java
+../src/main/java/core/Deadline.java
+../src/main/java/core/Todo.java
+../src/main/java/util/Storage.java
+../src/main/java/util/Parser.java
+../src/main/java/util/TronaldDumpException.java
+../src/main/java/commands/EventCommand.java
+../src/main/java/commands/DeleteCommand.java
+../src/main/java/commands/UnmarkCommand.java
+../src/main/java/commands/ListCommand.java
+../src/main/java/commands/Command.java
+../src/main/java/commands/TodoCommand.java
+../src/main/java/commands/MarkCommand.java
+../src/main/java/commands/DeadlineCommand.java
+../src/main/java/commands/ExitCommand.java
+../src/main/java/commands/CommandFactory.java
